requirements.txt
fastapi==0.115.12
uvicorn==0.34.1
ruff==0.11.5
pydantic-settings==2.8.1
passlib==1.7.4
PyJWT==2.10.1
httpx==0.28.1
pytest==8.3.5
pytest-asyncio==0.26.0
pytest-cov==6.1.1